move public functions and their tests to a separate file
