+++ ecs/world_test.go
@@ -113,25 +113,6 @@ func TestWorldNewEntites(t *testing.T) {
 	}
 }
 
-func TestWorldIDs(t *testing.T) {
-	w := NewWorld()
-
-	posID := ComponentID[Position](&w)
-	rotID := ComponentID[rotation](&w)
-
-	res1ID := ResourceID[Position](&w)
-	res2ID := ResourceID[Velocity](&w)
-
-	assert.Equal(t, uint8(0), posID.id)
-	assert.Equal(t, uint8(1), rotID.id)
-
-	assert.Equal(t, uint8(0), res1ID.id)
-	assert.Equal(t, uint8(1), res2ID.id)
-
-	assert.Equal(t, []ID{id(0), id(1)}, ComponentIDs(&w))
-	assert.Equal(t, []ResID{{id: 0}, {id: 1}}, ResourceIDs(&w))
-}
-
 func TestWorldComponents(t *testing.T) {
 	w := NewWorld()
 
@@ -200,30 +181,6 @@ func TestWorldComponents(t *testing.T) {
 	w.RemoveEntity(e0)
 	assert.Panics(t, func() { w.Has(newEntityGen(1, 0), posID) })
 	assert.Panics(t, func() { w.Get(newEntityGen(1, 0), posID) })
-}
-
-func TestWorldComponentInfo(t *testing.T) {
-	w := NewWorld()
-	_ = ComponentID[Velocity](&w)
-	posID := ComponentID[Position](&w)
-
-	info, ok := ComponentInfo(&w, posID)
-	assert.True(t, ok)
-	assert.Equal(t, info.Type, reflect.TypeOf(Position{}))
-
-	info, ok = ComponentInfo(&w, ID{id: 3})
-	assert.False(t, ok)
-	assert.Equal(t, info, CompInfo{})
-
-	resID := ResourceID[Velocity](&w)
-
-	tp, ok := ResourceType(&w, resID)
-	assert.True(t, ok)
-	assert.Equal(t, tp, reflect.TypeOf(Velocity{}))
-
-	tp, ok = ResourceType(&w, ResID{id: 3})
-	assert.False(t, ok)
-	assert.Equal(t, tp, nil)
 }
 
 func TestWorldIds(t *testing.T) {
@@ -1421,15 +1378,6 @@ func TestWorldResources(t *testing.T) {
 	assert.Panics(t, func() { w.Resources().Remove(rotID) })
 }
 
-func TestRegisterComponents(t *testing.T) {
-	world := NewWorld()
-
-	ComponentID[Position](&world)
-
-	assert.Equal(t, id(0), ComponentID[Position](&world))
-	assert.Equal(t, id(1), ComponentID[rotation](&world))
-}
-
 func TestWorldBatchRemove(t *testing.T) {
 	world := NewWorld()
 
@@ -1587,10 +1535,6 @@ func TestArchetypeGraph(t *testing.T) {
 
 	archEmpty4 := world.findOrCreateArchetype(arch012, []ID{}, []ID{posID, rotID, velID}, Entity{})
 	assert.Equal(t, archEmpty, archEmpty4)
-}
-
-type withSlice struct {
-	Slice []int
 }
 
 func TestWorldRemoveGC(t *testing.T) {
